enhance info endpoint and bump to `v0.1.13` (#117)
In a previous PR some(much desired) version information was added to the
`/info` endpoint, however in certain environments(our docker container)
this blows up.

This PR prevents the endpoint from blowing up, while also providing some
optionality and defaults to fall back on and bumps version to `v0.1.13`

Added the text below to the `README.md` file:

- `server` is read from the `process.env.npm_package_name` variable that
`npm` provides. If that does not exist, it will check for a
`DWN_SERVER_PACKAGE_NAME` environment variable set by the user, or
otherwise it will default to `@web5/dwn-server`.
- `version` and `sdkVersion` are read from the `package.json` file. It
will locate the file's path either from the
`process.env.npm_package_json` variable that `npm` provides. If that
does not exist, it will check for a `DWN_SERVER_PACKAGE_JSON`
environment variable set by the user, or otherwise it will default to
`/dwn-server/package.json` which is the path within the default Docker
container build.